+++ pkg/providers/cps/testdata/TestResDVValidation/create_validation_with_timeout.tf
@@ -6,9 +6,8 @@ resource "akamai_cps_dv_validation" "dv_validation" {
   enrollment_id = 1
   sans = [
     "san.test.akamai.com",
-    "san2.test.akamai.com",
   ]
   timeouts {
-    default = "1h"
+    default = "2ms"
   }
 }