+++ pkg/providers/edgeworkers/testdata/TestResEdgeWorkersEdgeWorker/edgeworker_lifecycle/edgeworker_create_with_timeout.tf
@@ -1,0 +1,13 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+resource "akamai_edgeworker" "edgeworker" {
+  name             = "example"
+  group_id         = "grp_12345"
+  resource_tier_id = 54321
+  local_bundle     = "testdata/TestResEdgeWorkersEdgeWorker/bundles/bundleForCreate.tgz"
+  timeouts {
+    default = "2h"
+  }
+}